put a dent in compiler warnings

+++ Source/Engine/GraphicsEngine.cpp
@@ -130,7 +130,7 @@ bool GraphicsEngine::getDisplayAdapters() {
 	return true;
 }
 
-bool GraphicsEngine::initializeAdapter(int adapterIndex) {
+bool GraphicsEngine::initializeAdapter(size_t adapterIndex) {
 
 	selectedAdapterIndex = adapterIndex;
 	selectedAdapter = adapters[selectedAdapterIndex];
@@ -228,8 +228,8 @@ void GraphicsEngine::initializeViewport() {
 	ZeroMemory(&mainViewport, sizeof(D3D11_VIEWPORT));
 	mainViewport.TopLeftX = 0;
 	mainViewport.TopLeftY = 0;
-	mainViewport.Width = Globals::WINDOW_WIDTH;
-	mainViewport.Height = Globals::WINDOW_HEIGHT;
+	mainViewport.Width = (float) Globals::WINDOW_WIDTH;
+	mainViewport.Height = (float) Globals::WINDOW_HEIGHT;
 	mainViewport.MinDepth = 0.0f;
 	mainViewport.MaxDepth = 1.0f;
 	deviceContext->RSSetViewports(1, &mainViewport);
